made numberFormatting folder, use roundAndAddCommas for formatting
