+++ src/components/priceAndTodaysPriceChange/index.js
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux';
 
 import PriceAndTodaysPriceChange from './PriceAndTodaysPriceChange';
-import { addCommas } from '../../utils/formatting/dataItem';
+import { roundAndAddCommas } from '../../utils/formatting/numberFormatting';
 
 import {
   LABEL_CURRENT_PRICE,
@@ -23,22 +23,22 @@ const calculatePriceChangePercentage = stockData =>
   calculatePriceChange(stockData) / getCurrentPrice(stockData) * 100;
 
 const getFormattedCurrentPrice = stockData =>
-  addCommas(
-    getCurrentPrice(
-      stockData
-    ).toFixed(VALUE_PRECISION_CURRENT_PRICE)
+  roundAndAddCommas(
+    getCurrentPrice(stockData),
+    VALUE_PRECISION_CURRENT_PRICE
   );
 
 const getFormattedPriceChange = stockData =>
-  addCommas(
-    calculatePriceChange(
-      stockData
-    ).toFixed(VALUE_PRECISION_PRICE_CHANGE)
+  roundAndAddCommas(
+    calculatePriceChange(stockData),
+    VALUE_PRECISION_PRICE_CHANGE
   );
 
 const getFormattedPriceChangePercentage = stockData =>
-  `${calculatePriceChangePercentage(stockData)
-    .toFixed(VALUE_PRECISION_PRICE_CHANGE_PERCENTAGE)}%`;
+  `${roundAndAddCommas(
+    calculatePriceChangePercentage(stockData),
+    VALUE_PRECISION_PRICE_CHANGE_PERCENTAGE
+  )}%`;
 
 const getStockData = state => {
   const symbol = state.selectedStock;
