+++ src/utils/formatting/numberFormatting.js
@@ -1,0 +1,17 @@
+/* Takes a number or string */
+const addCommas = number => {
+  let parts = number.toString().split('.');
+
+  const wholeNumberIndex = 0;
+  const wholeNumber = parts[wholeNumberIndex];
+  const wholeNumberWithCommas = wholeNumber.replace(/\B(?=(\d{3})+(?!\d))/g, ',');
+
+  parts[wholeNumberIndex] = wholeNumberWithCommas;
+
+  return parts.join('.');
+};
+
+export const roundAndAddCommas = (value, precision) => {
+  let result = value.toFixed(precision);
+  return addCommas(result);
+};
